[impl] #3 Split monolith into installable library + CLI.
- Library package src/claude_busy_monitor/ with _core.py (types,
  probing, transcript, public API) and _cli.py (ANSI rendering +
  main()). __init__.py re-exports the public API and reads
  __version__ via importlib.metadata.
- CHANGES.md is the version source of truth; pyproject reads it via
  hatch dynamic regex.
- Initial release v0.1.0: first usable CLI + stable public surface.
- Makefile gains an `install` target that runs `uv tool install --force .`
  (puts claude-busy-monitor on ~/.local/bin/).
- README-STATE-DETECTION.md companion path updated to point at _core.py.
- Original monolith claude_busy_monitor.py deleted.

Closes #3

Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
